initial table changes
