+++ hackathon_site/dashboard/frontend/src/components/dashboard/ItemTable/ItemTable.tsx
@@ -39,6 +39,8 @@ import {
 import {
     GeneralOrderTitle,
     GeneralOrderTableTitle,
+    GeneralPendingTable,
+    GeneralReturnTable,
 } from "components/general/OrderTables/OrderTables";
 
 export const CheckedOutTables = () =>
@@ -194,125 +196,24 @@ export const ReturnedTable = () => {
     const dispatch = useDispatch();
     const orders = useSelector(returnedOrdersSelector);
     const fetchOrdersError = useSelector(orderErrorSelector);
-    const hardware = useSelector(hardwareSelectors.selectEntities);
     const isVisible = useSelector(isReturnedTableVisibleSelector);
     const toggleVisibility = () => dispatch(toggleReturnedTable());
 
     return (
-        <Container
-            className={styles.tableContainer}
-            maxWidth={false}
-            disableGutters={true}
-        >
-            <GeneralOrderTitle
-                {...{
-                    title: "Returned items",
-                    isVisible,
-                    toggleVisibility,
-                }}
-            />
-
-            {isVisible &&
-                (!orders.length || fetchOrdersError ? (
-                    <Paper elevation={2} className={styles.empty} square={true}>
-                        {fetchOrdersError
-                            ? `Unable to view returned items.`
-                            : "Please bring items to the tech table and a tech team member will assist you."}
-                    </Paper>
-                ) : (
-                    orders.map((order) => (
-                        <div key={order.id}>
-                            <GeneralOrderTableTitle orderId={order.id} />
-                            <TableContainer
-                                component={Paper}
-                                elevation={2}
-                                square={true}
-                            >
-                                <Table
-                                    className={styles.table}
-                                    size="small"
-                                    aria-label="returned table"
-                                >
-                                    <TableHead>
-                                        <TableRow>
-                                            <TableCell className={styles.widthFixed} />
-                                            <TableCell
-                                                className={styles.width6}
-                                                align="left"
-                                            >
-                                                Name
-                                            </TableCell>
-                                            <TableCell
-                                                className={styles.widthFixed}
-                                                align="right"
-                                            >
-                                                Qty
-                                            </TableCell>
-                                            <TableCell
-                                                className={styles.width4}
-                                                align="right"
-                                            >
-                                                Time
-                                            </TableCell>
-                                            <TableCell
-                                                align="left"
-                                                className={styles.width2}
-                                            >
-                                                Condition
-                                            </TableCell>
-                                        </TableRow>
-                                    </TableHead>
-                                    <TableBody>
-                                        {order.hardwareInOrder.map((row) => (
-                                            <TableRow key={row.id}>
-                                                <TableCell align="left">
-                                                    <img
-                                                        className={styles.itemImg}
-                                                        src={
-                                                            hardware[row.hardware_id]
-                                                                ?.picture ??
-                                                            hardwareImagePlaceholder
-                                                        }
-                                                        alt={
-                                                            hardware[row.hardware_id]
-                                                                ?.name
-                                                        }
-                                                    />
-                                                </TableCell>
-                                                <TableCell align="left">
-                                                    {hardware[row.hardware_id]?.name}
-                                                </TableCell>
-                                                <TableCell align="right">
-                                                    {row.quantity}
-                                                </TableCell>
-                                                <TableCell
-                                                    align="right"
-                                                    className={styles.noWrap}
-                                                >
-                                                    {row.time}
-                                                </TableCell>
-                                                <TableCell
-                                                    align="left"
-                                                    className={styles.noWrap}
-                                                >
-                                                    {row.part_returned_health}
-                                                </TableCell>
-                                            </TableRow>
-                                        ))}
-                                    </TableBody>
-                                </Table>
-                            </TableContainer>
-                        </div>
-                    ))
-                ))}
-        </Container>
+        <GeneralReturnTable
+            {...{
+                orders,
+                fetchOrdersError,
+                isVisible,
+                toggleVisibility,
+            }}
+        />
     );
 };
 
 export const PendingTables = () => {
     const dispatch = useDispatch();
     const orders = useSelector(pendingOrderSelectors.selectAll);
-    const hardware = useSelector(hardwareSelectors.selectEntities);
     const isVisible = useSelector(isPendingTableVisibleSelector);
     const isCancelOrderLoading = useSelector(cancelOrderLoadingSelector);
     const toggleVisibility = () => dispatch(togglePendingTable());
@@ -341,68 +242,7 @@ export const PendingTables = () => {
                         key={pendingOrder.id}
                         data-testid={`pending-order-table-${pendingOrder.id}`}
                     >
-                        <GeneralOrderTableTitle
-                            orderId={pendingOrder.id}
-                            orderStatus={pendingOrder.status}
-                        />
-                        <TableContainer component={Paper} elevation={2} square={true}>
-                            <Table
-                                className={styles.table}
-                                size="small"
-                                aria-label="pending table"
-                            >
-                                <TableHead>
-                                    <TableRow>
-                                        <TableCell className={styles.widthFixed} />
-                                        <TableCell
-                                            className={styles.width6}
-                                            align="left"
-                                        >
-                                            Name
-                                        </TableCell>
-                                        <TableCell
-                                            className={`${styles.width1} ${styles.noWrap}`}
-                                            align="right"
-                                        >
-                                            Requested Qty
-                                        </TableCell>
-                                        <TableCell
-                                            className={`${styles.width1} ${styles.noWrap}`}
-                                            align="right"
-                                        >
-                                            Granted Qty
-                                        </TableCell>
-                                        <TableCell className={styles.widthBuffer} />
-                                    </TableRow>
-                                </TableHead>
-                                <TableBody>
-                                    {pendingOrder.hardwareInTableRow.map((row) => (
-                                        <TableRow key={row.id}>
-                                            <TableCell align="left">
-                                                <img
-                                                    className={styles.itemImg}
-                                                    src={
-                                                        hardware[row.id]?.picture ??
-                                                        hardwareImagePlaceholder
-                                                    }
-                                                    alt={hardware[row.id]?.name}
-                                                />
-                                            </TableCell>
-                                            <TableCell align="left">
-                                                {hardware[row.id]?.name}
-                                            </TableCell>
-                                            <TableCell align="right">
-                                                {row.quantityRequested}
-                                            </TableCell>
-                                            <TableCell align="right">
-                                                {row.quantityGranted}
-                                            </TableCell>
-                                            <TableCell className={styles.widthBuffer} />
-                                        </TableRow>
-                                    ))}
-                                </TableBody>
-                            </Table>
-                        </TableContainer>
+                        <GeneralPendingTable {...{ pendingOrder }} />
                         <div
                             style={{
                                 display: "flex",
