+++ hackathon_site/dashboard/frontend/src/pages/TeamDetail/TeamDetail.module.scss
@@ -1,0 +1,3 @@
+.dividerMargin {
+    margin-top: 20px;
+}
